+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/service/telqos/AlarmCommand.java
@@ -1,78 +1,78 @@
 package com.dwarfeng.bitalarm.impl.service.telqos;
 
 import com.dwarfeng.bitalarm.stack.service.AlarmQosService;
+import com.dwarfeng.springtelqos.node.config.TelqosCommand;
 import com.dwarfeng.springtelqos.sdk.command.CliCommand;
 import com.dwarfeng.springtelqos.stack.command.Context;
 import com.dwarfeng.springtelqos.stack.exception.TelqosException;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.apache.commons.lang3.tuple.Pair;
-import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
-@Component
+@TelqosCommand
 public class AlarmCommand extends CliCommand {
+
+    private static final String COMMAND_OPTION_ONLINE = "online";
+    private static final String COMMAND_OPTION_OFFLINE = "offline";
+
+    private static final String[] COMMAND_OPTION_ARRAY = new String[]{
+            COMMAND_OPTION_ONLINE,
+            COMMAND_OPTION_OFFLINE
+    };
 
     private static final String IDENTITY = "alarm";
     private static final String DESCRIPTION = "报警功能上线/下线";
-    private static final String CMD_LINE_SYNTAX_ONLINE = "alarm -online";
-    private static final String CMD_LINE_SYNTAX_OFFLINE = "alarm -offline";
-    private static final String CMD_LINE_SYNTAX = CMD_LINE_SYNTAX_ONLINE + System.lineSeparator() +
-            CMD_LINE_SYNTAX_OFFLINE;
+    private static final String CMD_LINE_SYNTAX_ONLINE = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_ONLINE);
+    private static final String CMD_LINE_SYNTAX_OFFLINE = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_OFFLINE);
+
+    private static final String[] CMD_LINE_ARRAY = new String[]{
+            CMD_LINE_SYNTAX_ONLINE,
+            CMD_LINE_SYNTAX_OFFLINE
+    };
+
+    private static final String CMD_LINE_SYNTAX = CommandUtil.syntax(CMD_LINE_ARRAY);
+
+    private final AlarmQosService alarmQosService;
 
     public AlarmCommand(AlarmQosService alarmQosService) {
         super(IDENTITY, DESCRIPTION, CMD_LINE_SYNTAX);
         this.alarmQosService = alarmQosService;
     }
 
-    private final AlarmQosService alarmQosService;
-
     @Override
     protected List<Option> buildOptions() {
         List<Option> list = new ArrayList<>();
-        list.add(Option.builder("online").optionalArg(true).hasArg(true).desc("上线服务").build());
-        list.add(Option.builder("offline").optionalArg(true).hasArg(true).desc("下线服务").build());
+        list.add(Option.builder(COMMAND_OPTION_ONLINE).desc("上线服务").build());
+        list.add(Option.builder(COMMAND_OPTION_OFFLINE).desc("下线服务").build());
         return list;
     }
 
     @Override
     protected void executeWithCmd(Context context, CommandLine cmd) throws TelqosException {
         try {
-            Pair<String, Integer> pair = analyseCommand(cmd);
+            Pair<String, Integer> pair = CommandUtil.analyseCommand(cmd, COMMAND_OPTION_ARRAY);
             if (pair.getRight() != 1) {
-                context.sendMessage("下列选项必须且只能含有一个: -online -offline");
+                context.sendMessage(CommandUtil.optionMismatchMessage(COMMAND_OPTION_ARRAY));
                 context.sendMessage(CMD_LINE_SYNTAX);
                 return;
             }
             switch (pair.getLeft()) {
-                case "online":
+                case COMMAND_OPTION_ONLINE:
                     alarmQosService.startAlarm();
-                    context.sendMessage("记录功能已上线!");
+                    context.sendMessage("报警功能已上线!");
                     break;
-                case "offline":
+                case COMMAND_OPTION_OFFLINE:
                     alarmQosService.stopAlarm();
-                    context.sendMessage("记录功能已下线!");
+                    context.sendMessage("报警功能已下线!");
                     break;
             }
         } catch (Exception e) {
             throw new TelqosException(e);
         }
-    }
-
-    @SuppressWarnings("DuplicatedCode")
-    private Pair<String, Integer> analyseCommand(CommandLine cmd) {
-        int i = 0;
-        String subCmd = null;
-        if (cmd.hasOption("online")) {
-            i++;
-            subCmd = "online";
-        }
-        if (cmd.hasOption("offline")) {
-            i++;
-            subCmd = "offline";
-        }
-        return Pair.of(subCmd, i);
     }
 }